+++ NEWS.md
@@ -1,3 +1,13 @@
+# dirichletprocess 0.2.1
+
+* Added AppVeyor, Travis-CI and codecov.io badges.
+* Added penalised log-likelihood step for posterior cluster parameter inference.
+* Added exponential mixture model `DirichletProcessExponential`. 
+* Updated `plot`. Multivariate Gaussian models can now be plotted.
+* Various bug fixes.
+* Updated description.
+ 
+
 # dirichletprocess 0.2.0
 
 * First public release.